Add headless support join variable quick fix. Fixes #641 (#642)
* Add headless support join variable quick fix. Fixes #641
